requirements.txt
Flask>=2.0.0
Werkzeug>=2.0.0
Jinja2>=3.0.0
itsdangerous>=2.0.0
click>=8.0.0
MarkupSafe>=2.0.0
gunicorn>=20.1.0

# 科学计算库
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=0.24.0
Pillow>=8.0.0

# 基本依赖
matplotlib>=3.5.0
requests>=2.25.0
python-dateutil>=2.8.2
six>=1.16.0
joblib>=1.0.0

# 可视化和工具库
pytz>=2021.1
certifi>=2021.5.30
pyparsing>=2.4.7
chardet>=4.0.0
idna>=2.10
threadpoolctl>=2.1.0